Updated m6anet to 2.0.2
With the updated biocontainer for m6anet being built, can update
the singularity and docker container with the latest version.

This version changed how m6anet is called.  It has `m6anet` and
the analysis as a parameter (`inference`, `dataprep`) instead of
`m6anet-run_inference` and `m6anet-dataprep` commands.

There are now two files for output from `m6anet inference`, but
the existing output of '*' captures those.  In the future, might
be nice to specify these individually and emit them as separate
outputs for downstream analysis.